requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask_wtf==1.2.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
matplotlib==3.8.0
plotly==5.17.0
bokeh==3.2.2
joblib==1.3.2
gunicorn==21.2.0
pytest==7.4.2
black==23.9.1
isort==5.12.0
flake8==6.1.0